requirement.txt
openai-agents
pyyaml
pydantic
langfuse
python-dotenv 
